ignore/scripts/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agents-framework"
version = "0.1.0"
description = "A minimal, extensible framework for LLM agents"
readme = "README.md"
authors = [
    {name = "Kevin Lappe", email = "kevin@averageintelligence.ai"}
]
license = {text = "MIT"}
requires-python = ">=3.8"
keywords = ["agents", "llm", "ai", "framework"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

# Minimal dependencies for quick start
dependencies = [
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "httpx>=0.24",
    "rich>=13.0",  # For nice console output
]

# Optional dependencies for advanced features
[project.optional-dependencies]
anthropic = ["anthropic>=0.18"]
openai = ["openai>=1.0"]
google = ["google-generativeai>=0.3"]
local = ["ollama>=0.1"]
vectorstore = [
    "chromadb>=0.4",
    "langchain>=0.1",
]
mcp = ["mcp>=0.1"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
]
all = [
    "agents-framework[anthropic,openai,google,local,vectorstore,mcp,dev]"
]

[project.urls]
Homepage = "https://github.com/klappe/agents"
Documentation = "https://github.com/klappe/agents/blob/main/README.md"
Repository = "https://github.com/klappe/agents.git"
Issues = "https://github.com/klappe/agents/issues"

[project.scripts]
agents = "agents_runtime.cli:main"
agent-run = "agents_runtime.runner:run"

[tool.setuptools]
packages = ["agents_runtime", "agents"]

[tool.setuptools.package-data]
agents = ["**/*.yaml", "**/*.json", "**/*.toml", "**/*.md"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
